+++ vork/src/main/scala/vork/internal/io/FileWatcher.scala
@@ -1,7 +1,7 @@
 package vork.internal.io
 
-import io.methvin.watcher.DirectoryChangeEvent.EventType
 import io.methvin.watcher.DirectoryChangeEvent
+import io.methvin.watcher.DirectoryChangeEvent.EventType
 import io.methvin.watcher.DirectoryChangeListener
 import io.methvin.watcher.DirectoryWatcher
 import java.nio.file.Files
