+++ vork/src/main/scala/vork/internal/pos/BinarySearch.scala
@@ -1,4 +1,4 @@
-package vork.internal.markdown
+package vork.internal.pos
 
 import scala.annotation.tailrec
